FbApi & Notepad: pulled viewHolderConfigurer out of RVAdapter and made as own class
